backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9.9,<3.0
gunicorn>=21.2.0,<22.0
whitenoise>=6.6.0,<7.0
argon2-cffi>=23.1.0,<24.0
bcrypt>=4.2.0,<5.0
dj-database-url>=2.1.0,<3.0
